refactor: move ActivityRepository to repositories package
