+++ src/main/java/com/alkemy/ong/domain/activity/ActivityDefaultGateway.java
@@ -1,7 +1,7 @@
 package com.alkemy.ong.domain.activity;
 
 import com.alkemy.ong.data.entities.ActivityEntity;
-import com.alkemy.ong.data.repository.ActivityRepository;
+import com.alkemy.ong.data.repositories.ActivityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
